Fallback to regular data bag item loading
This could be used for testing purposes where a cookbook project is
run under test-kitchen and there are data bag item fixtures set up.
When the Chef Vault fails to load, we safely fall back to loading a
data bag item of the same name from the bag, presuming this is for
test purposes.