[COOK-4591] - implement chef_vault_secret resource
This commit introduces the chef_vault_secret resource, with
test-kitchen convergence support and post-convergence tests using
ServerSpec.

In order to avoid awkward user creation that doesn't have a private
key Chef Vault can use the public key for the secret, a special
node/client is used (hydroelectric). The private key used here was
generated so there is actually valid data such that Chef Vault works.
This is a valid RSA private key, but it shouldn't be used anywhere
outside testing this cookbook, as publishing it makes it immediately
not secure for any purpose. You probably shouldn't use it for testing,
but it's okay :-).

* add a suite to .kitchen.yml
* add the provider and resource as libraries subclassing LWRPBase
* create a test cookbook recipe to setup everything required to test
  that the resource will create an encrypted data bag item that uses
  ChefVault:
  - install build-essential (required for cheffish dependency)
  - install cheffish
  - create the data bag using cheffish (chef_data_bag)