update for serverspec 2
The command gets \ escaped in ways that result in the command not being
found for some reason. Moving the “cd” to the command itself instead of
using :pre_command seems to fix the issue in this test.

Also need to update the helper for the new way of initializing
serverspec, and don’t pin a version of cheffish since the current
release has the code we need.