Fix compile_time property for Chef 12 AND 11
TL;DR, when using chef-sugar, it brings in a compile_time method to all
Chef::Resource classes, including ::ChefGem. This is ok on Chef 12.1,
because the instance method will be used (thanks Ruby!). This is not ok
on Chef 11 (and maybe 12.0), because compile_time from chef-sugar takes
a block parameter, and will cause an ArgumentError exception.

By checking that the instance methods (not those from ancestors)
includes compile_time, we can safely use the property on chef_gem.