Allow databag fallbag to be disabled globally
This provides an attribute node['chef-vault']['databag_fallback'] that will
prevent the behavior where an unencrypted data bag is loaded if it is present
instead of a vault item. If your policy requires it, this can be set in a role
or environment to force loading only from vault items.